Added the combined `index.html`
This is self-contained single file with a lot of in-line code (all the plots, images, styles). It is not meant for dev but instead when submitting a single HTML is allowed, this can be used.
Some size optimization was done as the Plotly JS scripts are only copied once (with the first iframe). Skeleton loaders were removed as they kept blinking behind the iframes.